+++ src/ytextractor/core/core.nim
@@ -1,9 +1,13 @@
 ## Core functions and constants for lib
+when defined js: import std/asyncjs
+else: import std/asyncdispatch
 
 from std/strutils import find, replace, multiReplace, join, split
 from std/json import parseJson, JsonNode, newJObject, items, hasKey, `{}`, getStr, keys
+from std/tables import `[]`
 
 from pkg/util/forStr import between
+import pkg/unifetch
 
 import ytextractor/core/cookies
 import ytextractor/core/types; export types
@@ -14,60 +18,37 @@ proc parseYoutubeJson*(
   ## Parses the html to get `ytInitialPlayerResponse` and `ytInitialData`
   try:
     result.ytInitialData = parseJson html.
-      between("var ytInitialData = ", "};", catchAll = true)[19..^2]
+      between("var ytInitialData = ", "};") & "}"
     result.ytInitialPlayerResponse = parseJson html.
-      between("var ytInitialPlayerResponse = ", "};", catchAll = true)[29..^2]
+      between("var ytInitialPlayerResponse = ", "};") & "}"
   except: discard
 
-const userAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/87.0.4280.141 Safari/537.36"
+const userAgent = "Mozilla/5.0 (Windows NT 10.0; rv:109.0) Gecko/20100101 Firefox/115.0"
 
-when not defined js:
-  from std/httpclient import newHttpClient, newHttpHeaders, get, code, Http200,
-                             body, close, `==`
-  import std/httpclient except `[]`
-  from std/tables import `[]`
-
-  proc fetch*(url: string): string =
-    ## Fetches a page using `GET` method and returns the response body
-    ## It also uses a browser user agent to not be blocked as bot
-    var
-      cookies = getCookies()
-      client = newHttpClient()
-    client.headers = newHttpHeaders({
-      "accept-language": "en-US,en;q=0.9",
-      "user-agent": userAgent,
-      "referer": url,
-      "dpr": "1",
-      "dnt": "1",
-      "cookie": $cookies,
-    })
-    let res = client.get url
-    block cookieExtract:
-      if res.headers.hasKey "set-cookie":
-        for setCookie in res.headers.table["set-cookie"]:
-          let cookie = setCookie[0..setCookie.find(';') - 1].split '='
-          cookies[cookie[0]] = cookie[1]
-        cookies.save()
-    if res.code == Http200:
-      result = res.body
-    client.close()
-else:
-  when defined nodejs:
-    # TODO nodejs implementation
-    discard
-  else:
-    from pkg/ajax import newXmlHttpRequest, setRequestHeader, open, send
-    proc fetch*(url: string): string =
-      ## Fetches a page using `GET` method and returns the response body
-      ## It also uses a browser user agent to not be blocked as bot
-      ## JS backend
-      var req = newXmlHttpRequest()
-      req.open("GET", url, async = false)
-      req.setRequestHeader("accept-language", "en-US,en;q=0.9")
-      # req.setRequestHeader("user-agent", userAgent) # Refused to set unsafe header
-      req.send()
-      if req.status == 200:
-        result = $req.responseText
+proc fetch*(url: string): Future[string] {.async.} =
+  ## Fetches a page using `GET` method and returns the response body
+  ## It also uses a browser user agent to not be blocked as bot
+  var
+    cookies = getCookies()
+    client = newUniClient()
+  client.headers = newHttpHeaders({
+    "accept-language": "en-US,en;q=0.9",
+    "user-agent": userAgent,
+    "referer": url,
+    "dpr": "1",
+    "dnt": "1",
+    "cookie": $cookies,
+  })
+  let res = await client.get url
+  block cookieExtract:
+    if res.headers.hasKey "set-cookie":
+      for setCookie in res.headers.table["set-cookie"]:
+        let cookie = setCookie[0..setCookie.find(';') - 1].split '='
+        cookies[cookie[0]] = cookie[1]
+      save cookies
+  if res.code == Http200:
+    result = res.body
+  close client
 
 func findInJson*(nodes: JsonNode; key: string): JsonNode =
   ## Gets an json array and returns a object